Fix decimal displaying rule for configured difficulty rating values.

+++ src/SudokuStudio/Configuration/TechniqueData.cs
@@ -12,4 +12,4 @@ namespace SudokuStudio.Configuration;
 /// one of the such difficulty levels cannot be modified its difficulty level and rating value.
 /// </para>
 /// </param>
-public readonly record struct TechniqueData(decimal Rating, DifficultyLevel Level);
+public readonly record struct TechniqueData(int Rating, DifficultyLevel Level);
